requirements.txt
pandas==1.3.0
plotly==4.14.3
streamlit==1.0.0
cloudscraper==1.2.58
Pillow==9.0.0
requests==2.27.1